requirements.txt
asgiref==3.8.1
autopep8==2.0.0
certifi==2023.11.17
distlib==0.3.4
Django==4.2.11
djangorestframework==3.15.1
filelock==3.6.0
platformdirs==2.5.2
pycodestyle==2.10.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
virtualenv==20.14.1
